build: set up for easy install with poetry and pip
