+++ pyproject.toml
@@ -1,0 +1,21 @@
+[tool.poetry]
+name = "basicmemnet"
+version = "0.1.0"
+description = "minimal memnet implementation"
+authors = ["Your Name <you@example.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+matplotlib = "3.7.5"
+networkx = "3.1"
+pydot = "2.0.0"
+nltk = "3.8.1"
+bson = "0.5.10"
+
+[tool.poetry.group.dev.dependencies]
+black = "24.4.2"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
